fix(frontend routes): fixed URL in move.js
