+++ .eslintrc.json
@@ -6,5 +6,8 @@
         "ecmaFeatures": {
             "jsx": true
         }
+    },
+    "react": {
+        "prop-types": 0
     }
 }
