+++ browser/routes/move.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
 
 export const movePlayer = (playerId, newPosition, possibleMoves) => {
-  axios.put(`/api/${playerId}/move`, { newPosition, possibleMoves })
+  axios.put(`/api/game/gameOne/player/${playerId}/move`, { newPosition, possibleMoves })
 };
